+++ .gitignore
@@ -31,3 +31,4 @@ src/braket/default_simulator/.DS_Store
 test/.DS_Store
 test/unit_tests/.DS_Store
 test/unit_tests/braket/.DS_Store
+.DS_Store
